Physics quick updates and Button Press soundfx
Button Presses now have sound effects, and I changed the physics on the rock a little, so it would fall faster an feel less floaty. I don't think it needed it, necessarily, but it's just one variable if we want to change it back, and I think it adds some.